% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{league_url_data}
\alias{league_url_data}
\title{All the available soccer leagues and tournaments with ESPN scoreboard urls}
\format{A data frame with one row for each league/tournament and two variables:
\describe{
 \item{name}{the name of the league or tournament}
 \item{url}{scoreboard url for the league/tournament}
}}
\usage{
league_url_data
}
\description{
All the available soccer leagues and tournaments with ESPN scoreboard urls
}
\details{
The league/tournament names and respective urls are strings.
}
\examples{
data(league_url_data)

}
\keyword{datasets}
