% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_commentary.R
\name{scrape_commentary}
\alias{scrape_commentary}
\title{Scrapes ESPN match commentary for a given game ID}
\usage{
scrape_commentary(game_id)
}
\arguments{
\item{game_id}{ESPN game ID to scrape the match commentary for.}
}
\value{
Data frame of the match commentary with a row for each event in the
commentary along with the columns described above.
}
\description{
Scrapes ESPN match commentary for a given game ID
}
\details{
Through string parsing of the commentary the following columns are
created:
\itemize{
 \item{"game_id"} - Unique identifier for the game from ESPN
 \item{"commentary"} - Raw text commentary
 \item{"match_time"} - Match time provided by ESPN in character format
 \item{"team_one"} - Name of first team in game based on order on game page
 \item{"team_two"} - Name of second team in game based on order on game page
 \item{"team_one_score"} - Score for team one
 \item{"team_two_score"} - Score for team two
 \item{"half_begins"} - Indicator for when a half begins
 \item{"half_end"} - Indicator for when a half ends
 \item{"match_end"} - Indicator for when the match ends
 \item{"shot_attempt"} - Indicator for if a shot was attempted
 \item{"penalty_shot"} - Indicator for if a the shot attempt was a penalty shot
 \item{"shot_result"} - Outcome for shot attempt, either: goal, blocked,
 missed, saved, or own_goal
 \item{"shot_with"} -  Which foot (left or right) or header person shot with
 \item{"shot_where"} - Location from field relative to box shot took place from
 \item{"shot_by_player"} - Player who attempted shot
 \item{"shot_by_team"} - Team that attempted shot
 \item{"assist_by_player"} - Player who assisted shot attempt
 \item{"assist_type"} - Denotes if assist was with a through or cross
 \item{"foul"} - Indicator if there was a foul
 \item{"foul_by_player"} - Which player committed the foul
 \item{"foul_by_team"} - Which team committed the foul
 \item{"net_location"} - Location relative to net of shot attempt
 \item{"follow_set_piece"} - Indicator if followed set piece situation
 \item{"follow_corner"} - Indicator if followed corner kick
 \item{"offside"} - Indicator if offside was called
 \item{"offside_team"} - Which team was offside
 \item{"offside_player"} - Which player was offside
 \item{"offside_pass_from"} - Which player passed the ball leading to offside
 \item{"shown_card"} - Indicator if a card was shown
 \item{"card_type"} - Type of card: red or yellow
 \item{"card_player"} - Player given the card
 \item{"card_team"} - Team given the card
 \item{"video_review"} - Indicator if a video review took place
 \item{"video_review_event"} - Which event was reviewed
 \item{"video_review_result"} - Result of the video review
 \item{"delay_in_match"} - Indicator for if a delay in match was called
 \item{"delay_team"} - Which team caused the delay
 \item{"free_kick_won"} - Indicator for if a free kick was won
 \item{"free_kick_player"} - Which player won the free kick
 \item{"free_kick_team"} - Which team won the free kick
 \item{"free_kick_where"} - Where was the free kick won
 \item{"corner"} - Indicator if a corner kick was awarded
 \item{"corner_team"} - Which team was awarded the corner kick
 \item{"corner_conceded_by"} - Which player conceded the corner kick
 \item{"substitution"} - Indicator for if a substitution took place
 \item{"sub_injury"} - Indicator for if substitution took place because of injury
 \item{"sub_team"} - Which team had the substitution
 \item{"sub_player"} - Which player came in during the substitution
 \item{"replaced_player"} - Which player was replaced during the substitution
 \item{"penalty"} - Indicator if penalty occurred
 \item{"team_drew_penalty"} - Which team was awarded the penalty
 \item{"team_conceded_penalty"} - Which team conceded the penalty
 \item{"player_drew_penalty"} - Which player drew the penalty
 \item{"player_conceded_penalty"} - Which player conceded the penalty
 \item{"half"} - Numeric indicator for the half
 \item{"comment_id"} - Row id for comment
 \item{"match_time_numeric"} - Numeric value for match time minutes
}
}
\examples{
# Get the match commentary for Serbia vs Costa Rica in the 2018 World Cup:
srb_crc_commentary <- scrape_commentary(498194)
}
