% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_game_ids.R
\name{scrape_scoreboard_ids}
\alias{scrape_scoreboard_ids}
\title{Scrapes the game ids from the current soccer scoreboard on ESPN}
\usage{
scrape_scoreboard_ids()
}
\value{
Vector of game ids from the current day's soccer scoreboard
}
\description{
Scrapes the game ids from the current soccer scoreboard on ESPN
}
\examples{
# Scrape today's game ids:
today_game_ids <- scrape_scoreboard_ids()
}
